requirements.txt
aiofiles==0.4.0
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
httptools==0.0.13
idna==2.8
multidict==4.5.2
numpy==1.16.3
pandas==0.24.2
pymongo==3.4.0
python-dateutil==2.8.0
python-dotenv==0.10.2
pytz==2019.1
sanic==19.3.1
six==1.12.0
ujson==1.35
urllib3==1.25.2
uvloop==0.12.2
websockets==6.0
yarl==1.3.0
sanic-openapi==0.5.3
psycopg2==2.8.3
sanic-jwt==1.3.2
cryptography==2.7